Add tests for several modules
