+++ test/team_test.rb
@@ -5,10 +5,10 @@ class TeamTest < Minitest::Test
   def setup
     @team = Team.new({
             team_id: 1,
-            franchiseId: 23,
-            teamName: "Atlanta United",
+            franchiseid: 23,
+            teamname: "Atlanta United",
             abbreviation: "ATL",
-            Stadium: "Mercedes-Benz Stadium",
+            stadium: "Mercedes-Benz Stadium",
             link: "/api/v1/team/1"
      })
   end
